Bugfix for keyboard-navigasjon etter arbeidsflate-bytte
