+++ src/komponenter/header/meny/ekspanderbar-meny/desktop-visningsmeny/DesktopVisningsmeny.less
@@ -23,13 +23,12 @@
             max-height: 60rem;
             max-width: 100%;
             visibility: visible;
-            box-shadow: 11px 8px 8px rgba(0, 0, 0, 0.4);
-            @supports (filter: drop-shadow(0 0)) {
-                filter: drop-shadow(11px 8px 8px rgba(0, 0, 0, 0.4));
-                box-shadow: none;
-            }
-
-            transform: none;
+            //box-shadow: 11px 8px 8px rgba(0, 0, 0, 0.4);
+            //@supports (filter: drop-shadow(0 0)) {
+            //    filter: drop-shadow(11px 8px 8px rgba(0, 0, 0, 0.4));
+            //    box-shadow: none;
+            //}
+            //transform: none;
         }
     }
