+++ src/komponenter/header/meny/ekspanderbar-meny/desktop-visningsmeny/meny-uinnlogget/MenyUinnlogget.less
@@ -1,7 +1,7 @@
 @import (reference) '../../../../../../styling-variabler.less';
 
 .meny {
-  @side-margin: 0.5em;
+  @side-margin: 1.5em;
 
   .lenke-stil() {
     text-decoration: underline;
@@ -30,7 +30,7 @@
   &__topp-seksjon {
     margin: 1.5rem -3rem;
     padding: 1rem 3rem;
-    background-color: rgba(200, 200, 200, 0.4);
+    //background-color: rgba(200, 200, 200, 0.4);
   }
 
   &__topp-seksjon-lenke {
