+++ src/komponenter/header/meny/ekspanderbar-meny/desktop-visningsmeny/meny-uinnlogget/bunn-seksjon/BunnSeksjon.tsx
@@ -4,27 +4,15 @@ import BunnseksjonLenke from './BunnseksjonLenke';
 import KbNav, { NaviGroup } from '../../keyboard-navigation/kb-navigation';
 import { MenuValue } from '../../../../../../../utils/meny-storage-utils';
 import { Language } from '../../../../../../../reducer/language-duck';
-import { AppState } from '../../../../../../../reducer/reducer';
-import { connect } from 'react-redux';
 import { finnTekst } from '../../../../../../../tekster/finn-tekst';
 import { bunnLenker } from './BunnseksjonLenkedata';
-import { Dispatch } from '../../../../../../../redux/dispatch-type';
-import { finnArbeidsflate } from '../../../../../../../reducer/arbeidsflate-duck';
 
-interface OwnProps {
+interface Props {
     classname: string;
-}
-
-interface StateProps {
+    settArbeidsflateFunc: () => void;
     arbeidsflate: MenuValue;
     language: Language;
 }
-
-interface DispatchProps {
-    settArbeidsflateFunc: () => void;
-}
-
-type Props = OwnProps & StateProps & DispatchProps;
 
 const BunnSeksjon = ({ classname, language, arbeidsflate, settArbeidsflateFunc }: Props) => {
     const cls = BEMHelper(classname);
@@ -50,13 +38,4 @@ const BunnSeksjon = ({ classname, language, arbeidsflate, settArbeidsflateFunc }
     );
 };
 
-const mapStateToProps = (state: AppState): StateProps => ({
-    arbeidsflate: state.arbeidsflate.status,
-    language: state.language.language
-});
-
-const mapDispatchToProps = (dispatch: Dispatch): DispatchProps => ({
-    settArbeidsflateFunc: () => dispatch(finnArbeidsflate()),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(BunnSeksjon);
+export default BunnSeksjon;
